+++ src/Aws/CloudFront/CloudFrontClient.php
@@ -64,7 +64,7 @@ use Guzzle\Service\Resource\ResourceIteratorInterface;
  */
 class CloudFrontClient extends AbstractClient
 {
-    const LATEST_API_VERSION = '2014-01-31';
+    const LATEST_API_VERSION = '2014-05-31';
 
     /**
      * Factory method to create a new Amazon CloudFront client using an array of configuration options.