+++ tests/Aws/Tests/CloudSearch/CloudSearchClientTest.php
@@ -14,9 +14,10 @@
  * permissions and limitations under the License.
  */
 
-namespace Aws\Tests\CloudWatch;
+namespace Aws\Tests\CloudSearch;
 
 use Aws\CloudSearch\CloudSearchClient;
+use Guzzle\Service\Resource\Model;
 
 class CloudSearchClientTest extends \Guzzle\Tests\GuzzleTestCase
 {
@@ -34,5 +35,21 @@ class CloudSearchClientTest extends \Guzzle\Tests\GuzzleTestCase
         $this->assertInstanceOf('Aws\Common\Signature\SignatureV4', $this->readAttribute($client, 'signature'));
         $this->assertInstanceOf('Aws\Common\Credentials\Credentials', $client->getCredentials());
         $this->assertEquals('https://cloudsearch.us-east-1.amazonaws.com', $client->getBaseUrl());
+    }
+
+    public function testCanGetDomainClient()
+    {
+        $client = $this->getMockBuilder('Aws\CloudSearch\CloudSearchClient')
+            ->disableOriginalConstructor()
+            ->setMethods(array('describeDomains'))
+            ->getMock();
+        $client->expects($this->any())
+            ->method('describeDomains')
+            ->will($this->returnValue(new Model(array(
+                'DomainStatusList' => array(array('SearchService' => array('Endpoint' => 'foo.cloudsearch.com')))
+            ))));
+
+        $domainClient = $client->getDomainClient('foo');
+        $this->assertEquals('https://foo.cloudsearch.com', $domainClient->getBaseUrl());
     }
 }